requirements.txt
pytest==7.1.2
pytest-selenium
termcolor
allure-python-commons
allure-pytest